Move cli.py into streamlit.web
